% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_to_epsg.r
\name{prj_to_epsg}
\alias{prj_to_epsg}
\title{Title}
\usage{
prj_to_epsg(prj)
}
\arguments{
\item{prj}{}
}
\value{
CRS
}
\description{
Title
}
\examples{
\dontrun{
##  https://stat.ethz.ch/pipermail/r-sig-geo/2016-January/023861.html
prj <- paste0(readLines("110m_admin_1_states_provinces_shp.prj"))
prj_to_epsg(prj)
## CRS arguments:
##  +init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84
## +towgs84=0,0,0

prj_1 <- 'PROJCS["Transverse_Mercator",
    GEOGCS["GCS_OSGB 1936",
    DATUM["D_OSGB_1936",
    SPHEROID["Airy_1830",6377563.396,299.3249646]],
    PRIMEM["Greenwich",0],UNIT["Degree",0.017453292519943295]],
    PROJECTION["Transverse_Mercator"],
    PARAMETER["latitude_of_origin",49],
    PARAMETER["central_meridian",-2],
    PARAMETER["scale_factor",0.9996012717],
    PARAMETER["false_easting",400000],
    PARAMETER["false_northing",-100000],
    UNIT["Meter",1]]'
prj_to_epsg(prj_1)
## CRS arguments:
##  +init=epsg:27700 +proj=tmerc +lat_0=49 +lon_0=-2 +k=0.9996012717
## +x_0=400000 +y_0=-100000 +datum=OSGB36 +units=m +no_defs +ellps=airy
## +towgs84=446.448,-125.157,542.060,0.1502,0.2470,0.8421,-20.4894
}
}
\author{
boB Rudis <bob at rudis.net>
}

